+++ Changelog.md
@@ -1,7 +1,10 @@
-# Changelog v1.7.2
+# Changelog v1.8
+
+## Additions
+- Added configuration to disable, move and resize all components
 
 ## Changes
-- Using buttons with key binds plays a sound now
+- Enchantments are now also included in the filter (including enchantment level)
 
 ## Fixes
-- Exception gets thrown on server side because of key mappings
+- Dimmed and highlighted slots are aligned incorrectly in shulker boxes